Make the value of RARRAY_LEN the same as on MRI.
We have a C function that back RARRAY_LEN but it was using size_t instead of
the long that MRI uses for the length field in the Array structure. The differenc
leads to needless warnings in C-extensions.

+++ vm/capi/19/include/ruby/ruby.h
@@ -872,8 +872,8 @@ VALUE rb_uint2big(unsigned long number);
   /** Remove and return first element of Array or nil. Changes other elements' indexes. */
   VALUE   rb_ary_shift(VALUE self);
 
-  /** Number of elements in given Array. @todo MRI specifies int return, problem? */
-  size_t  rb_ary_size(VALUE self);
+  /** Number of elements in given Array. */
+  long    rb_ary_size(VALUE self);
 
   /** Store object at given index. Supports negative indexes. Returns object. */
   void    rb_ary_store(VALUE self, long int index, VALUE object);
